refactor(emoji): extract add/replace form and share upload parsing
Pull the emoji add/replace form out of EmojiSearch into its own AddEmojiForm
component (EmojiSearch 466 to 249 lines). Extract shared upload parsing and
params schema into routes/emojis/parse.ts so create and replace routes stop
duplicating it. Replace the non-existent theme.Error/theme.Warning lookups
with explicit color constants, fixing a type error.
