+++ app/src/js/components/organisms/EmojiSearch.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useState } from "react";
 import styled from "styled-components";
 import { SearchBox } from "../atoms/SearchBox";
 import { type Emoji as EmojiType, EmojiDescriptor } from "../../types";
@@ -6,7 +6,7 @@ import { Icon } from "../atoms/Icon";
 import { ClassNames, cn } from "../../utils";
 import { Emoji } from "../molecules/Emoji";
 import { Button } from "../molecules/Button";
-import { FormHelpText, FormInput } from "../atoms/FormInput";
+import { AddEmojiForm } from "./AddEmojiForm";
 import { observer } from "mobx-react-lite";
 import { useApp } from "../contexts/appState";
 
@@ -72,79 +72,6 @@ export const EmojiSearchContainer = styled.div`
       .icon {
         margin: 0;
       }
-    }
-  }
-`;
-
-const AddEmojiForm = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  border-top: 1px solid ${(props) => props.theme.Strokes};
-  padding: 12px;
-
-  .add-emoji-row {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 8px;
-  }
-
-  .add-emoji-preview {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex: 0 0 48px;
-    width: 48px;
-    height: 48px;
-    border-radius: 8px;
-    border: 1px dashed ${(props) => props.theme.Strokes};
-    overflow: hidden;
-    cursor: pointer;
-    color: ${(props) => props.theme.Labels};
-    img {
-      width: 100%;
-      height: 100%;
-      object-fit: contain;
-    }
-  }
-
-  .add-emoji-shortname {
-    flex: 1 1 auto;
-    height: 40px;
-  }
-
-  .add-emoji-actions {
-    display: flex;
-    flex-direction: row;
-    gap: 8px;
-
-    button {
-      flex: 1 1 0;
-      padding: 11px 16px;
-    }
-  }
-
-  .add-emoji-error {
-    color: ${(props) => props.theme.Error ?? "#e5484d"};
-  }
-
-  .add-emoji-warning-panel {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 10px;
-    padding: 12px;
-    border-radius: 8px;
-    border: 1px solid ${(props) => props.theme.Warning ?? "#e5a23d"};
-    background-color: ${(props) => props.theme.Warning ?? "#e5a23d"}1a;
-    color: ${(props) => props.theme.Text};
-    font-size: 13px;
-    line-height: 18px;
-
-    .icon {
-      color: ${(props) => props.theme.Warning ?? "#e5a23d"};
-      flex: 0 0 auto;
     }
   }
 `;
@@ -230,86 +157,9 @@ export const EmojiSearch = observer(
     const [results, setResults] = useState<Record<string, EmojiType[]>>({});
     const [order, setOrder] = useState<string[]>([]);
     const [adding, setAdding] = useState(false);
-    const [pendingReplace, setPendingReplace] = useState(false);
-    const [shortname, setShortname] = useState("");
-    const [file, setFile] = useState<File | null>(null);
-    const [previewUrl, setPreviewUrl] = useState<string | null>(null);
-    const [error, setError] = useState<string | null>(null);
-    const [submitting, setSubmitting] = useState(false);
-    const fileInputRef = useRef<HTMLInputElement>(null);
     const app = useApp();
     const emojis = app.emojis.getAll();
     const fuse = app.emojis.getFuse();
-
-    useEffect(() => {
-      if (!file) {
-        setPreviewUrl(null);
-        return;
-      }
-      const url = URL.createObjectURL(file);
-      setPreviewUrl(url);
-      return () => URL.revokeObjectURL(url);
-    }, [file]);
-
-    useEffect(() => {
-      setPendingReplace(false);
-    }, [shortname, file]);
-
-    const resetForm = () => {
-      setAdding(false);
-      setPendingReplace(false);
-      setShortname("");
-      setFile(null);
-      setError(null);
-      setSubmitting(false);
-    };
-
-    const openAdd = () => {
-      resetForm();
-      setAdding(true);
-    };
-
-    const pickFile = (e: React.ChangeEvent<HTMLInputElement>) => {
-      const picked = e.target.files?.[0] ?? null;
-      if (picked && !picked.type.startsWith("image/")) {
-        setError("Please choose an image file");
-        return;
-      }
-      setError(null);
-      setFile(picked);
-    };
-
-    const submit = async () => {
-      const cleaned = shortname.trim().replace(/^:/, "").replace(/:$/, "");
-      if (!/^[a-zA-Z0-9_+-]+$/.test(cleaned)) {
-        setError("Invalid shortname. Use letters, numbers, _ + -");
-        return;
-      }
-      if (!file) {
-        setError("Please choose an image");
-        return;
-      }
-      const sn = `:${cleaned}:`;
-      const existing = app.emojis.get(sn);
-      const isCustom = Boolean(existing?.fileId) && !existing?.unicode;
-      if (isCustom && !pendingReplace) {
-        setPendingReplace(true);
-        return;
-      }
-      setSubmitting(true);
-      setError(null);
-      try {
-        if (isCustom) {
-          await app.emojis.replace(sn, file);
-        } else {
-          await app.emojis.create(sn, file);
-        }
-        resetForm();
-      } catch (err) {
-        setError(err instanceof Error ? err.message : "Failed to save emoji");
-        setSubmitting(false);
-      }
-    };
 
     useEffect(() => {
       if (!name || !fuse) {
@@ -384,76 +234,10 @@ export const EmojiSearch = observer(
           </div>
         </div>
         {adding
-          ? (
-            <AddEmojiForm className="cmp-add-emoji-form">
-              {pendingReplace
-                ? (
-                  <div className="add-emoji-warning-panel">
-                    <Icon icon="lock" size={16} />
-                    <span>
-                      {`:${
-                        shortname.trim().replace(/^:/, "").replace(/:$/, "")
-                      }: already exists. Replace it with the new image?`}
-                    </span>
-                  </div>
-                )
-                : (
-                  <>
-                    <div className="add-emoji-row">
-                      <div
-                        className="add-emoji-preview"
-                        onClick={() => fileInputRef.current?.click()}
-                      >
-                        {previewUrl
-                          ? <img src={previewUrl} alt="emoji preview" />
-                          : <Icon icon="icons" size={20} />}
-                      </div>
-                      <FormInput
-                        className="add-emoji-shortname"
-                        placeholder=":shortname:"
-                        value={shortname}
-                        autoFocus
-                        onChange={(e) => setShortname(e.target.value)}
-                        onKeyDown={(e) => {
-                          if (e.key === "Enter") submit();
-                        }}
-                      />
-                      <input
-                        ref={fileInputRef}
-                        type="file"
-                        accept="image/*"
-                        style={{ display: "none" }}
-                        onChange={pickFile}
-                      />
-                    </div>
-                    {error && (
-                      <FormHelpText className="add-emoji-error">
-                        {error}
-                      </FormHelpText>
-                    )}
-                  </>
-                )}
-              <div className="add-emoji-actions">
-                <Button
-                  type="secondary"
-                  onClick={() =>
-                    pendingReplace ? setPendingReplace(false) : resetForm()}
-                >
-                  {pendingReplace ? "Back" : "Cancel"}
-                </Button>
-                <Button type="primary" onClick={submit} disabled={submitting}>
-                  {submitting
-                    ? "Saving…"
-                    : pendingReplace
-                    ? "Replace"
-                    : "Add"}
-                </Button>
-              </div>
-            </AddEmojiForm>
-          )
+          ? <AddEmojiForm onClose={() => setAdding(false)} />
           : (
             <div className="add-emoji">
-              <Button type="secondary" onClick={openAdd}>
+              <Button type="secondary" onClick={() => setAdding(true)}>
                 ADD EMOJI
                 <Icon icon="plus" size={16} />
               </Button>
